Polled signal handling.
hbcxx now has lots of logic to ensure it always cleans up temporary files.
Switching to polled signal handling allows us to clean up even when we
exit with a signal pending.

The logic is made more complex because we still want sub-processes to
react normally to signals (complex compiles can be pretty slow and it
feels lumpy if they run with signals masked). Therefore we have to
drop in new implementations of system() to ensure signals are handled
properly.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>